Add all files to db, tag non media with non media tag.  This speeds up rescans so they dont have to rescan non media for media files
